test(note): migrate basic
