+++ core/src/components/note/test/basic/note.e2e.ts-snapshots/note-diff-ios-ltr-Mobile-Firefox-linux.png
