+++ core/src/components/note/test/basic/note.e2e.ts-snapshots/note-diff-md-ltr-Mobile-Safari-linux.png
