+++ core/src/components/note/test/basic/note.e2e.ts-snapshots/note-item-divider-end-md-ltr-Mobile-Chrome-linux.png
