+++ core/src/components/note/test/basic/note.e2e.ts-snapshots/note-item-divider-start-md-rtl-Mobile-Firefox-linux.png
