+++ core/src/components/note/test/basic/note.e2e.ts-snapshots/note-item-end-md-ltr-Mobile-Chrome-linux.png
