+++ core/src/components/note/test/basic/note.e2e.ts-snapshots/note-item-end-md-rtl-Mobile-Safari-linux.png
