+++ core/src/components/note/test/basic/note.e2e.ts-snapshots/note-item-start-md-ltr-Mobile-Firefox-linux.png
